Successfully implemented FOV + map revamp!
ARL.Fov is working now and enemies can move into and out of your field
of view.
Added a ton of new code to ARL.View to revamp how the map was drawn. Now
instead of drawing text into individual elements, we're drawing text
onto a canvas so you can't just highlight the page to see what's in the
dark. ^_^;;
Added some code to handle a buffer array for the DIRTY_LOCS array. Now I
can have multiple steps of a turn play out, queue all the things that
need to be redrawn, and push that queue to dirty all at once, so I'm not
doing multiple redraw operations through the course of a turn unless
it's deliberate.
Added code to do the initial draw of the map since we don't get that for
free anymore.
Added code to update the visibility of tiles before updating their glyph
and to ignore mobs standing on tiles that aren't within vision range
when deciding the current glyph.
Added a dedicated function to end the current turn so I can put cleanup
stuff in there to ensure it all gets called every turn.
Probably more stuff I'm not remembering.