Implemented ranged combat!
The player character can now fire on gobbos from range. The player's
ranged attack distance is equal to their Fov distance right now.
Pressing the F key will show a targeting cursor that will let you choose
a cardinal direction to fire an arrow in. Press Space to confirm your
choice and Escape to cancel the shot.
This update required additional View code to draw the cursor over the
map, as well as the implementation of input contexts to allow you to
move a targeting cursor without moving the player mob.
Some of this new code is rigid, but it will likely be revisited soon
when I implement gobbo archers. :D
I also added some new Util methods for generating lists of specific locs
in relation to a given loc that can be parsed for targeting.
I added a new tFireThru value to the TERRAIN_BASE for each terrain type,
as well as a tName value that I can grab for referring to a tile in
narration. The tFireThru value also opens up the possibility for things
like windows that you can see through but not walk through or fire
through or bars that you can see through and fire through but not walk
through. Adding those would require me to retool the mega-simple mob
pursuit AI though, so there's another project for down the road.