feat: a bunch of stuff I can't remember
I commit to this repo so infrequently that I can't honestly remember
some of the changes I've made since my last commit. It looks like my
last commit was the implementation of ranged combat so WHOOO BOY.

Once I have this committed, I'll do a follow-up commit with a clearer
indication of what was committed here.